SKILL-192: prove FULL validate repairs by receipts and confirmation identity closure.
Require a durable repair plan and per-identity substantiation before confirmation, and treat confirmation as closed only when discovery identities are absent from the confirmation finding set.
