+++ runtime-kotlin/runtime-application/src/main/kotlin/skillbill/application/featuretask/FeatureTaskRuntimePhasePromptComposer.kt
@@ -100,7 +100,7 @@ object FeatureTaskRuntimePhasePromptComposer {
       implementationContinuationDirective(briefing.phaseId, effectiveContinuation),
       retryCorrectionDirective(briefing, priorSchemaFailure, correctiveRepairContext),
       terminalRetryDirective(priorTerminalFailure),
-      outputContract(briefing, reviewPassNumber, priorBlockerFindingIds),
+      outputContract(briefing, reviewPassNumber, priorBlockerFindingIds, validationDepth, agentRunValidateFallback),
     ).filter(String::isNotBlank).joinToString(separator = "\n\n")
   }
 
@@ -354,6 +354,8 @@ object FeatureTaskRuntimePhasePromptComposer {
     briefing: FeatureTaskRuntimePhaseLaunchBriefing,
     reviewPassNumber: Int?,
     priorBlockerFindingIds: List<String>,
+    validationDepth: ValidationDepth,
+    agentRunValidateFallback: Boolean,
   ): String {
     val phaseId = briefing.phaseId
     return """
@@ -373,6 +375,8 @@ object FeatureTaskRuntimePhasePromptComposer {
       result for downstream phases (for example plan steps, changed files, findings, or
       validation results)${producedOutputsAddendum(
       briefing,
+      validationDepth,
+      agentRunValidateFallback,
     )}${dispositionAddendum(briefing, reviewPassNumber, priorBlockerFindingIds)}
     - "derived_notes": optional; when present, a non-empty string of notes for downstream
       phases
@@ -428,34 +432,14 @@ object FeatureTaskRuntimePhasePromptComposer {
   // thorough agent from delivering its verdict as a prose Markdown table the gate cannot read (and then
   // blocking after a blind retry loop). The two phase sets are disjoint, so at most one branch is ever
   // non-empty; every other phase returns "" so its output contract stays byte-for-byte unchanged.
-  private fun producedOutputsAddendum(briefing: FeatureTaskRuntimePhaseLaunchBriefing): String {
+  private fun producedOutputsAddendum(
+    briefing: FeatureTaskRuntimePhaseLaunchBriefing,
+    validationDepth: ValidationDepth,
+    agentRunValidateFallback: Boolean,
+  ): String {
     val phaseId = briefing.phaseId
     if (FeatureTaskRuntimePhaseWorkflowDefinition.isMutatingPhase(phaseId)) {
-      val remediation = if (briefing.auditRepairItemIds.isEmpty()) {
-        ""
-      } else {
-        "\n    - This is AUDIT-GAP REMEDIATION. produced_outputs MUST also include repair_item_results " +
-          "with exactly these ids in order: ${briefing.auditRepairItemIds.joinToString()}. Every result must " +
-          "contain only repair_item_id, outcome (fixed or already_satisfied), non-empty " +
-          "changed_paths_or_symbols, non-empty executed_verification, and structured result_evidence " +
-          "with observation, artifact_ref, and check_ref. observation MUST be exactly one of " +
-          "required_behavior_absent, verification_failed, contract_rejected, state_mismatch, fix_verified, " +
-          "already_satisfied_verified, resolution_verified, or recurrence_verified; use " +
-          "already_satisfied_verified when the behavior already existed and you verified it, and do not " +
-          "invent a synonym outside this list. artifact_ref MUST be a repository-relative path " +
-          "optionally followed by one :symbol; do not put a sentence, spaces, test description, command, " +
-          "result, or additional prose in artifact_ref. check_ref MUST be AC-###, F-###, or a single " +
-          "name ending in Test or Check (optionally followed by :symbol); do not put a command, result, " +
-          "sentence, spaces, or shell punctuation in check_ref.\n" +
-          "    - produced_outputs MUST include deferred_repair_item_ids. Completed remediation uses []; " +
-          "blocked remediation lists every remaining item, while unresolvable_repair identifies exactly one.\n" +
-          auditRemediationOutputExample(briefing.auditRepairItemIds)
-      }
-      return "\n    - produced_outputs MUST include a reconciliation report: a \"reconciled_state\" object\n" +
-        "      (or a \"reconciled_state\" entry) with \"reconciled\": true and concrete evidence that the\n" +
-        "      changed files are at their intended target state. A status of \"completed\" with the\n" +
-        "      reconciliation report missing or \"reconciled\" not true fails the schema gate loudly." +
-        FeatureTaskRuntimePhaseProjectionShapes.exampleFor(phaseId) + remediation
+      return mutatingProducedOutputsAddendum(briefing, validationDepth, agentRunValidateFallback)
     }
     val findings = FeatureTaskRuntimeVerificationSignalKeys.REVIEW_FINDINGS
     val verdict = FeatureTaskRuntimeVerificationSignalKeys.VERDICT
@@ -463,7 +447,11 @@ object FeatureTaskRuntimePhasePromptComposer {
       FeatureTaskRuntimePhaseWorkflowDefinition.PHASE_PREPLAN,
       FeatureTaskRuntimePhaseWorkflowDefinition.PHASE_PLAN,
       FeatureTaskRuntimePhaseWorkflowDefinition.PHASE_VALIDATE,
-      -> FeatureTaskRuntimePhaseProjectionShapes.exampleFor(phaseId)
+      -> FeatureTaskRuntimePhaseProjectionShapes.exampleFor(
+        phaseId,
+        validationDepth,
+        agentRunValidateFallback,
+      )
       FeatureTaskRuntimePhaseWorkflowDefinition.PHASE_REVIEW ->
         "\n    - This is a VERIFYING phase: produced_outputs MUST carry a \"$findings\" array (each entry a\n" +
           "      severity/message object; an explicit empty [] affirms no Blocker or Major findings) AND/OR a\n" +
@@ -486,6 +474,43 @@ object FeatureTaskRuntimePhasePromptComposer {
       )
       else -> ""
     }
+  }
+
+  private fun mutatingProducedOutputsAddendum(
+    briefing: FeatureTaskRuntimePhaseLaunchBriefing,
+    validationDepth: ValidationDepth,
+    agentRunValidateFallback: Boolean,
+  ): String {
+    val phaseId = briefing.phaseId
+    val remediation = if (briefing.auditRepairItemIds.isEmpty()) {
+      ""
+    } else {
+      "\n    - This is AUDIT-GAP REMEDIATION. produced_outputs MUST also include repair_item_results " +
+        "with exactly these ids in order: ${briefing.auditRepairItemIds.joinToString()}. Every result must " +
+        "contain only repair_item_id, outcome (fixed or already_satisfied), non-empty " +
+        "changed_paths_or_symbols, non-empty executed_verification, and structured result_evidence " +
+        "with observation, artifact_ref, and check_ref. observation MUST be exactly one of " +
+        "required_behavior_absent, verification_failed, contract_rejected, state_mismatch, fix_verified, " +
+        "already_satisfied_verified, resolution_verified, or recurrence_verified; use " +
+        "already_satisfied_verified when the behavior already existed and you verified it, and do not " +
+        "invent a synonym outside this list. artifact_ref MUST be a repository-relative path " +
+        "optionally followed by one :symbol; do not put a sentence, spaces, test description, command, " +
+        "result, or additional prose in artifact_ref. check_ref MUST be AC-###, F-###, or a single " +
+        "name ending in Test or Check (optionally followed by :symbol); do not put a command, result, " +
+        "sentence, spaces, or shell punctuation in check_ref.\n" +
+        "    - produced_outputs MUST include deferred_repair_item_ids. Completed remediation uses []; " +
+        "blocked remediation lists every remaining item, while unresolvable_repair identifies exactly one.\n" +
+        auditRemediationOutputExample(briefing.auditRepairItemIds)
+    }
+    return "\n    - produced_outputs MUST include a reconciliation report: a \"reconciled_state\" object\n" +
+      "      (or a \"reconciled_state\" entry) with \"reconciled\": true and concrete evidence that the\n" +
+      "      changed files are at their intended target state. A status of \"completed\" with the\n" +
+      "      reconciliation report missing or \"reconciled\" not true fails the schema gate loudly." +
+      FeatureTaskRuntimePhaseProjectionShapes.exampleFor(
+        phaseId,
+        validationDepth,
+        agentRunValidateFallback,
+      ) + remediation
   }
 
   /**
