+++ runtime-kotlin/runtime-application/src/main/kotlin/skillbill/application/featuretask/validation/model/ValidationGateCycleModels.kt
@@ -7,6 +7,8 @@ import skillbill.workflow.model.ValidationDepth
 import skillbill.workflow.taskruntime.model.FeatureTaskRuntimePhaseOutput
 import skillbill.workflow.taskruntime.model.FeatureTaskRuntimeValidationGateProgress
 import skillbill.workflow.taskruntime.model.FeatureTaskRuntimeValidationGateRunRecord
+import skillbill.workflow.taskruntime.model.FullValidateRepairPlanItem
+import skillbill.workflow.taskruntime.model.FullValidateSubstantiationReceipt
 import java.nio.file.Path
 
 sealed interface ValidationGateResolution {
@@ -80,6 +82,15 @@ data class ValidationGateCycleRequest(
   val changedPaths: List<String>,
   val repositoryCheckpoint: String,
   val agentRepairLauncher: ValidationGateAgentRepairLauncher,
-  /** Base ref for suppression-delta measurement; blank skips measurement with a loud block when markers exist. */
   val baseRef: String = "",
+)
+
+data class FullValidateRepairCoverageEvaluation(
+  val accepted: Boolean,
+  val reason: String,
+)
+
+data class FullValidateRepairParsedArtifacts(
+  val plan: List<FullValidateRepairPlanItem>?,
+  val receipts: List<FullValidateSubstantiationReceipt>,
 )
