+++ runtime-kotlin/runtime-application/src/test/kotlin/skillbill/application/featuretask/validation/FeatureTaskRuntimeSuppressionGateTest.kt
@@ -243,7 +243,33 @@ class FeatureTaskRuntimeSuppressionGateTest {
     } else {
       ScriptedGateRunner(listOf(passed(), passed(forced = true)))
     }
-    val git = object : WorkflowGitOperations by NoopWorkflowGitOperations, SuppressionEvidenceGitOperationsProvider {
+    val coordinator = FeatureTaskRuntimeValidationGateCoordinator(
+      resolver = declaredResolver(declaration),
+      runner = runner,
+      progressStore = ValidationGateProgressStore { _, p, _ -> progress += p },
+      suppressionDeltaService = FeatureTaskRuntimeSuppressionDeltaService(suppressionEvidenceGit(scenario)),
+      repoLocalConfig = defaultRepoLocalConfigPort(),
+    )
+    return coordinator.execute(
+      ValidationGateCycleRequest(
+        repoRoot = repoRoot,
+        request = minimalRequest(),
+        validationDepth = scenario.options.validationDepth,
+        changedPaths = listOf("src/Foo.kt"),
+        repositoryCheckpoint = "checkpoint",
+        baseRef = "base",
+        agentRepairLauncher = ValidationGateAgentRepairLauncher { findings, iteration ->
+          scenario.options.onRepairLaunch(findings, iteration)
+          completedSuppressionRepair(findings, scenario)
+        },
+      ),
+    )
+  }
+
+  private fun suppressionEvidenceGit(
+    scenario: SuppressionGateScenario,
+  ): skillbill.ports.workflow.WorkflowGitOperations =
+    object : WorkflowGitOperations by NoopWorkflowGitOperations, SuppressionEvidenceGitOperationsProvider {
       override val suppressionEvidenceOperations = object : SuppressionEvidenceGitOperations {
         override fun scopedPathContentsAgainstBase(
           repoRoot: Path,
@@ -262,60 +288,45 @@ class FeatureTaskRuntimeSuppressionGateTest {
         )
       }
     }
-    val coordinator = FeatureTaskRuntimeValidationGateCoordinator(
-      resolver = declaredResolver(declaration),
-      runner = runner,
-      progressStore = ValidationGateProgressStore { _, p, _ -> progress += p },
-      suppressionDeltaService = FeatureTaskRuntimeSuppressionDeltaService(git),
-      repoLocalConfig = defaultRepoLocalConfigPort(),
-    )
-    return coordinator.execute(
-      ValidationGateCycleRequest(
-        repoRoot = repoRoot,
-        request = minimalRequest(),
-        validationDepth = scenario.options.validationDepth,
-        changedPaths = listOf("src/Foo.kt"),
-        repositoryCheckpoint = "checkpoint",
-        baseRef = "base",
-        agentRepairLauncher = ValidationGateAgentRepairLauncher { findings, iteration ->
-          scenario.options.onRepairLaunch(findings, iteration)
-          val justificationOnly = findings.findings.isNotEmpty() &&
-            findings.findings.all {
-              it.ruleOrTestId == FeatureTaskRuntimeValidationGateCoordinator.SUPPRESSION_JUSTIFICATION_RULE_ID
-            }
-          val produced = mutableMapOf<String, Any?>()
-          if (!justificationOnly && findings.findings.isNotEmpty()) {
-            produced["validation_repair_plan"] = findings.findings.map { finding ->
-              mapOf("identities" to listOf(finding.identity()))
-            }
-            produced["substantiation_receipts"] = findings.findings.map { finding ->
-              mapOf(
-                "identity" to finding.identity(),
-                "root_cause" to "root ${finding.ruleOrTestId}",
-                "changed_paths_or_symbols" to listOf(finding.location ?: finding.ruleOrTestId),
-                "rationale" to "fixed ${finding.ruleOrTestId}",
-              )
-            }
-          }
-          if (scenario.options.repairJustifications.isNotEmpty()) {
-            produced["validation_result"] = mapOf(
-              "suppression_justifications" to scenario.options.repairJustifications,
-            )
-          }
-          val payload = JsonSupport.mapToJsonString(
-            mapOf(
-              "contract_version" to "0.3",
-              "phase_id" to "validate",
-              "status" to "completed",
-              "summary" to "repair",
-              "produced_outputs" to produced,
-            ),
-          )
-          ValidationGateAgentRepairResult.Completed(
-            FeatureTaskRuntimePhaseOutput(phaseId = "validate", iteration = 1, payload = payload),
-          )
-        },
+
+  private fun completedSuppressionRepair(
+    findings: ValidationFindingSetProjection,
+    scenario: SuppressionGateScenario,
+  ): ValidationGateAgentRepairResult {
+    val justificationOnly = findings.findings.isNotEmpty() &&
+      findings.findings.all {
+        it.ruleOrTestId == FeatureTaskRuntimeValidationGateCoordinator.SUPPRESSION_JUSTIFICATION_RULE_ID
+      }
+    val produced = mutableMapOf<String, Any?>()
+    if (!justificationOnly && findings.findings.isNotEmpty()) {
+      produced["validation_repair_plan"] = findings.findings.map { finding ->
+        mapOf("identities" to listOf(finding.identity()))
+      }
+      produced["substantiation_receipts"] = findings.findings.map { finding ->
+        mapOf(
+          "identity" to finding.identity(),
+          "root_cause" to "root ${finding.ruleOrTestId}",
+          "changed_paths_or_symbols" to listOf(finding.location ?: finding.ruleOrTestId),
+          "rationale" to "fixed ${finding.ruleOrTestId}",
+        )
+      }
+    }
+    if (scenario.options.repairJustifications.isNotEmpty()) {
+      produced["validation_result"] = mapOf(
+        "suppression_justifications" to scenario.options.repairJustifications,
+      )
+    }
+    val payload = JsonSupport.mapToJsonString(
+      mapOf(
+        "contract_version" to "0.3",
+        "phase_id" to "validate",
+        "status" to "completed",
+        "summary" to "repair",
+        "produced_outputs" to produced,
       ),
+    )
+    return ValidationGateAgentRepairResult.Completed(
+      FeatureTaskRuntimePhaseOutput(phaseId = "validate", iteration = 1, payload = payload),
     )
   }
